+++ render-demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @byrding/render-demo
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @byrding/core@0.3.0
+  - @byrding/react@0.1.4
+
 ## 0.0.3
 
 ### Patch Changes